Move ProcessIssues to own file
