+++ app/src/org/commcare/android/tasks/ZipTask.java
@@ -8,7 +8,7 @@ import org.commcare.android.database.SqlStorage;
 import org.commcare.android.database.user.models.FormRecord;
 import org.commcare.android.javarosa.AndroidLogger;
 import org.commcare.android.models.notifications.NotificationMessageFactory;
-import org.commcare.android.tasks.ProcessAndSendTask.ProcessIssues;
+import org.commcare.android.models.notifications.ProcessIssues;
 import org.commcare.android.tasks.templates.CommCareTask;
 import org.commcare.android.util.FileUtil;
 import org.commcare.android.util.FormUploadUtil;
